jvm: uninitialized value fields in a product type may be null
When cloning a value instance with value field instances that were not
initialized yet, maybe fields for local variables that were never used anyways,
these fields need to be checked if they are null before they are cloned.

A similar check existed already for unused fields for unused values in a choice
type, so this patch moves this functionality into a helper method.

This fixes the failures when running flang.dev's tests

  content/design/examples/monadiclifting_hello3.fz
  content/design/examples/monadiclifting_hello4.fz
  content/design/examples/monadiclifting_hello5.fz
  content/design/examples/monadiclifting_helloabort.fz