ast: Set type of `StrConst` to `String` to avoid confusion
I found it inconsistent to have a difference in the type of type inference and
for type checking, so I changed this back to return `String` just like
`typeForInferencing`.

To create instances of `Const_String` for these constants in later phases, there
used be be special handling in the interpreter, DFA, etc.  Now, instead, there
is a new method `AbsractConstant.typeOfConstant` that is redefined by `StrConst`
to return `Const_String`. This method is now used in later phases when the typeo
the value to be created for a constant is required.

I will extend the tutorial on flang.dev to explain this a little better.