ast: Suppress feature-not-found as subsequent error
Feature lookup in the current scope via Feature.findFieldDefInScope failed to
find the use of a field if that use was within a call that had an error such
that the arguments where no longer iterated.

Now, in case of an existing error in a call the current scope, a failed feature
lookup is no longer reported.

This removes additional errors that are currently reported by flang.dev's
examples/examples/man_or_boy2.fz.