ast: Fix endless recursion in case of recursive uses of args with inferred type
This example

  fib(a) => list a (fib a)

resulted in an endless recursion since the result type determination of `a`
required the result type of a call to `a`, which triggered result type
determination of `a` and so forth...

This patch detects recursive determination of result type of a Feature and
returns `void` in this case.  Furthermore, the result type of a call is not
overwritten by `void` for a call to an argument field with inferred type.