jvm: Fixed result of CodeGen.combinePreconditionAndCall for unit and void type
The code that created a Pair was wrong in this case. Since this is fairly tricky
and required at another location as well, I added a helper method `makePair`.

Also performed small cleanup for handling of outer ref in tail calls that can
now use the newly added `JVM.javaTypeOfTarget` method.

This fixes loads for test failures in the previous commit.