ast/lib/tests: Cleanup in UnresolvedType.findGenerics
The goal of this work is to unify `UnresolvedType.findGenerics` and
`UnresolvedType.resolveFeature` into one single method that is performed in one
pass.  This is a first, but big step that modifies `findGeneric` to use
`lookupType`.

`AbstractFeature.getGeneric` has been removed.  New methods
`AbstractFeature.generic` and `AbstractFeature.genericType` are added, but not
as a replacement, but as a first steop to remove class `Generic`.

`findGenerics` now requires a `Resolution` argument, which hence has been added
to a couple of methods that end up calling `findGenerics`.
`Feature.findGenerics` was moved to `Resolution.findGenerics` for the same
reason.

There are several semantic changes:

- Equally named type parameters in inner and outer features are still permitted,
  but they cause an error when used since the use is ambiguous. Disambiguation
  using `x.this.T` does not work yet.  Base library's `net.channel.replace.T`
  had to be renamed to avoid this ambiguity.

- type parameters are no longer inherited, so there is no need and it is no
  longer possible to `redef` an inherited type parameter that is just passed
  through to the parent. Base library features `cons`, `Effect_Call`,
  `mutatle.array`, `Mutable_Array`, `oneway_monad`, `stream`,
  `container.hash_map`, `container.Linked_list`,
  `container.Mutable_Linked_list`, `container.orderd_map`, `container.ps_map`,
  `container.sorted_array`, `fuzion.java.Array` were changed accordingly.

- visibility of type parameters is now taken into account, which is why
  `array(|2|3).T` and `Sequence.T` are now marked public (and some other type
  parameters in the base library should be made pulbic as well).

- Redundant (and conflicting) type parameters `cache_item.T`,
  `list.type.fmat.A`, `list.type.concat_monoid.A` were removed.

Tests were added to check that ambiguous type parameter accesses now result in
an error (`tests/generics_negative`) and one test output was updated due to
errors reported in a different order now.  Two tests had `redef` in front of
type parameters that had to be removed.