air: WIP: replace Clazz.actualClazz by Clazzes.down(), fix #2214
The problem with `Clazz.actualClazz` was that it replaces all generics found in
parents and in outer features.  This is a problem when a feature inherits from
one of its outer features since there are two possible values of the actual
generics and for the examplein #2214, the wrong one was used.

With this patch, `Clazzes.down` now hands down an inherted type along the
inheritance change in each nesting level of features, so we have much better
control how inherited types are replaced by the actual values provided in the
inherts-call.

This patch also adds a test case with the code from #2214 plus the flang.dev
example `primes6_broken_map.fz` that showed this problem since a long time.