ast: Fix check if there is an ambiguous partial application
The problem was that for a call without an explicit target, Call._target is
originally null and the called feature was searched in all outer features. If
found, _target is set to the corresponding `outerFeature.this` expression.

THe code to check that there is no ambiguous partial application now must not
restrict its search to this target, but use the null target.

In case of fields, using a specific target could even result in finding more
fields due to scope checking that only happens for unqualified (i.e., w/o
target) calls to fields.

This caused a failure in flang.dev's design/examples/aborteffect.fz.