lib: Redefine count and drop for array backed sequences
The redefinition of `count` in `Mutable_Array` now has O(1) performance.

Similarly, redefine `Sequence.drop` in `array` and result of `array.slice`. This
improves the performance of `drop n` from O(n) to O(1).

I did not add a redefinition of `drop` to `Mutable_Array`, I am not sure if
`Mutable_Array` should be a `Sequence` at all.