ast: Fix type inference from actual args in case actuals get replaced
This fixes a bug that was revealed by automatically calling features with open
type parameters (eg., t.0 instead of t.values.0 in tuples): When the actual
expression gets replaced by something else, the type inference still used the
original expression resulting in potentially wrong types being inferred.

This patch no longer stores the actual argument that was found, but instead
stores the call that was found.  To determine the actual type, the actual is
taken form the call resulting in the new actual in case it was replaced during
resolveTypes or syntax sugar phases.

This also cleans up the somewhat clumsy check to prevent infinite recursion and
uses a simple flag instead.