lib/be: fix failures in tests/local_mutate and tests/stdin
There are two problems:

First, tests/local_mutate checks that mutating a locally mutable variable
outside of the scope of the local mutate effect by catching the resulting panic,
so the previous change to use a precondition instead of panic breaks this test.
Using panic, however, adds the panic effect to small examples like HelloWorld,
which is not desirable. So this patch implements a workaround that enables the
use for fuzion.std.panic in the base lib.

Second, intrinsics may currently not have preconditions, so I had to split up
`effect.abort` into one feature with precondition that calls a second `abort0`
that is intrinsic.