me: Replace runtime clazzes by clazz-local map of actual clazzes, fix #2408
The use of unique ids for all AST elements that need runtime clazz information
plus arrays large enough to hold all these ideas in every clazz instance results
in memory demand quadratic to the code size resulting in the high memory demand
for arrays used in ryu.fz and reported in #2408.

This patch replaces these large arrays by per-clazz TreeMaps that contain only
those actual clazzes that are actually needed for that clazz.

This also adds some NYI-comments related to #2411 and #2412.

This removes caching of Callable created in the interpreter, so this might have
a negative impact on interpreter performance.  We might to re-introduce a
caching mechanism if this turns out to be very costly.