me: Replace `clazz`/`innerClazz`/`runtimeClazz` in AST by Clazz.actualClazzes
There were two problems here, first one is architectural: AST should not contain
fields relevant for the later phases. Second, having only one clazz will fail as
soon as a constant's type depends on actual type parameters of any surrounding
features. E.g., an `option T` might have an `as_array` feature that uses an
empty array constant to return for the `nil` option.  The clazz of the type of
this constant will be different for `option i32` and `option bool`.

This patch uses `Clazz.saveActualClazzes()` to store the clazz of a constant. To
be able to do this late during FUIR, `AbstractConstant` now contains an `origin`
method to get the original `AbstractCall` or `InlineArray` instance to obtain
the clazz of the result type.

Also did some further cleanup:

 - Moved code to visit the code of an `InlineArray` from `Clazzes.findClazzes`
   to `InlineArray.java`.

 - FUIR.constData can no longer be called on a Call or InlineArray (it probably
   never was).

 - FUIR.toStack() was simplified slightly for AbstractCall and InlineArray, in
   particular an unused, constant InlineArray will now be dumped completely.