doc/util/fe: First intergration of rules, error ids and reference manual
Rules of the Fuzion language may now be embedded in the source code, they will
be wrapped by rule.adoc when generating the reference manual.

Rulese have an ID that consists of a domain and detail in that domain that is
addressed by the rule.

Errors now may (and eventually will) have an ID that corresponds to the rule
that is broken when this error is created.  Errors for the rule SRCF_UTF8 have
been changed to use their ID for error reporting.

Preliminary infrastructure to document the Unicode version Id in the reference
manual was implemented.