jvm: Fix race when printing warning statistics
The problem shows occasionally in Jenkins for tests like
design/example/arg_intri.fz, the problem is as follows: The compiler creates
several warnings in the middle end for missing intrinsics, and these warnings
turn into a fatal error at runtime.

Since the error is produced in a new thread, the fatal error output followed by
exit(1) races with the warning statistics output during shutdown of the main
thread resulting in the warning statistics occasionally not being printed.

To solve this, the statistics are printed before the compiled code is executed
and the main thread is terminated quietly.