fix: "no targets for access"
fixes #2231
fixes #2273
fixes #2280

Essentially the problem was/is that the DFA finds that the lambda passed to `go` will never return. This leads to a push of `null`/`void` in the AbstractInterpreter, stopping any further processing of this code block.
But the backends do not know that this calls never returns and continue processing the code block. For the C-backend this works by accident, I think.
It did however break the JVM-Backend.

To reiterate the situation is that the JVM-backend generates code the DFA did not analyze. This is problemetic since
`accessedClazzes` did not contain a call to outer ref `#^Effect_Call.call` which led to the bugs described in the issues.
```
ex  is
  eff : simple_effect is
    something is
      abort

  _ := eff.go ()->
    ex.eff.env.something
```