lib: more links, comments and code for composition
Added links to Gabriel Lebec talk and Devid Kenan and Chris Rathman pages.

Added birds for identity: idiot / ibis

Added M/Mockingbird \f.ff, which is tough in a statically typed language, cannot
be defined in Haskell and probably neither in Fuzion, but still playing around
with it.  Also added M* wich seems to be definable in JavaScript, need to check
if that helps.

Added thrush and vireo.

Moved before1/before2 closer to compose since they are all Bluebirds.

Added Y and Z combinator, but mostly as comments for now.