examples/javaGraphics: fix call to java.lang.Thread.sleep
Java 21 introduced a new method java.lang.Thread.sleep that takes a
java.time.Duration in addition to the variant that takes a long. The
fzjava mangling chooses this as a default variant now, making these
calls lead to unexpected type errors. Fix this by explicitly selecting
the version of the sleep method that takes a long.