be/jvm: For -jar and -classes backends, fix arguments and include cmd as args[0]
The script created for -jar and -classes did not propagate the command name as
part of the arguments, so the first argument was falsely taken as the command
name.

With this patch, the command name is passed explicitly via a property
`fuzion.command`, which will be set to the script file in case of -jar/-classes
targets and to the `fz` command in case of the `-jvm` target.