parser: conditions are now syntactically the same as `exprs`, fix #149
Syntactially, the conditions in a pre/post condition or the (unused) invariant
are now parsed as `exprs`, i.e. like an indented block in a feature declaration.

Conditions are no longer separated by commas, but by semicolons or non-indenting
new lines.

The Fuzion grammar becomes slightly simpler with rules `cond` and `condList`
removed.

Updated syntax of conditions in base lib and tests where `,` was used, added
`debug:` qualifiers at some places and `(`/`)` for clarity, it is a bit
unfortunate that, e.g., `debug: -a = b` is parsed as `((debug.postfix :) - a) =
b`...