lib: spawn processes
Initial implementation for spawning processes with arguments and custom environment variables (all backends/all supported plattforms) Also reading stdout/stderr, writing to stdin of spawned processes as well as waiting for the process to finish and getting its exit code.
`infix |` operator is included for shell-like piping from one process to the next.

Limitiations:
Currently the piping of data is done manually in Fuzion in a background thread.
Furthermore spawning of processes is not threadsafe yet.

On Java this uses ProcessBuilder+InputStream+OutputStream
On C/POSIX posix_spawn+pipes are used
On C/Win CreateProcess+CreatePipe are used