lib/tests: Relax precondition for `step` when creating `interval`
For an interval `a..a:s`, it is now allowed that step is any value including
`s=0`, which gives the interval with singe element `a`.

This fixes `content/examples/examples/fib_iterative_neg.fz` from
fuzion-lang.dev, which gives an example for why this relaxation makes sense.

This also required changes in `size` and `contains`.

This patch also improves `interval.as_string` for infinite intervals and adds
`postifx ..∞` as an alias for `postifx ..`.

Also added a test to check various combinations of values for `from`, `through`
and `step`.