ir/fuir/mir: Add support for `site` instead of `c`/`i` for codeblock and index
A `site` is an int the identifies an expression in a code block, so it is a more
conveninient way compared ot passing the pair `c`/`i` that is so far used in
FUIR.  Eventually, codeblock indices should be removed completely and replaced
by `site`s only.

This patch adds a convenience method `IR.addCode` that manages the calculation
of `site` and `getExpr` as a temporary convenience method to get an Expr for
given `c`/`i`.