dfa: improve error produced for uninitialized field
Got this error before intrinsics for Java interface during DFA initialized the
`msg` field of an `error` result properly and it would have helped to get more
detail, which this patch provides.

Also slightly changed the output for missing effect errors and restructured the
handling of Contexts.

Removed code related to a HACK that no longer exists.