dfa: Distinguish initial and actual effect values.
Initial effect values are the values that are used for a call to use/go/run that
installs this effect. Environments are distinguished by these values.

Actual effect values are the initial values joined with all values an effect
might get replaced with via a call to instrinsic `effect.replace`.