fix dynamic_type used on features defined inside of type features
fixes #2402

This failed:
`say $"{3}".dynamic_type`
because under the hood this uses `String.type.concat` which returns an anonymous feature `ref: String` which is defined inside of the type feature `String.type.concat`.

For features inside of type features no type feature was added which led to `AbstractType.typetype()` returning just the type of `Type`. This triggered the error in the later stages.

Now type features are added even for feature defined inside of type features.

I had to change name of `type.concat` to `type.concat1` since type features are now generated for type.concat(0 args) and type.concat(1 arg) leading to "Duplicate feature declaration" since
we get type.concat(THIS#TYPE) and type.concat(THIS#TYPE, 1 arg)

Another solutions to the problem would be to dissallow calling dynamic_type on feature defined inside of type feature and raise an error in this case.