ast: change loop scoping semantics
previously this worked:
```
  for i := 10 do
  say i
```
but this does not work:
```
  for x in 1..10
        i := 10
  do
  say i
```

This patch changes the semantics of the loop in a way that the first example also results in a feature not found error.

I think this makes loop scoping more intuitive.

We loose the ability to access any non-iterating index variables until the first iterating index variable.