parser: move parsing of `.env` and `.type` to `callTail`
This avoids speculative parsings to distinguish a call `a.b.c` from an env
`a.b.env` or a type `a.b.type` and simplifies the parsers.

Since with this change, a call` is no longer always an instance of `Call`, added
a new rule `pureCall` that returns only a call and is used, e.g., in inherits
clauses.

This reduces the time to build java.base.fum by about 10% (from 21s down to 19s
on my machine).