air: Implicitly instantiate type features
This solves a tricky problem with idim87ex.fz: Here, `effect.is_installed` is
called, which is an intrinsic constructor defined in a type feature that results
in a bool. Since the type feature's clazz was not marked as instantiated,
Clazz.isInstantiated() for `effect.is_installed` also returned `false` and the
choice generics TRUE and FALSE of the bool result did hance not get marked as
instantiated.

Finally, since in `idiom87ex.fz`, there was no other instantiation of values
`TRUE` or `FALSE`, no code was created for any match on the bool result of
`isInstantiated`, resulting in the default `exit` effect not being installed.