lib/parser: Add `fuzion.runtime.fault` and support `check` expressions
`fuzion.runtime.fault` is an effect that is supposed to become the effect
triggered by a failed runtime check such as in a pre- or post-conditions.

As a first test for this effect, this patch implements the `check` expression,
which is just syntax sugar for an `if` checking a condition and calling
`check_fault` in case that fails.