ast/tests: Nicer types in error messages
First, produce a nicer error message for an expression in a contract that is not
boolean, include the actual type and highlight the expected type.

Since the actual type in `tests/reg_issu1838` is the type of a type parameter,
also improve the output generated for types of type parameters, in particular,
use `Any.type` or `String.type` instead of `Any.#type#1 Any` or `String.#type#1
String`.