lib/parser/tests: `pre`/`post` conditions in redef now need `else`/`then`
As a first step towards properly inheriting pre- and postconditions, this commit
changes the grammar and fron end to require that redefined features use `pre
else` to relax a precondition and `post then` to strenghten a postcondition.

This patch does not change the semantics, i.e., inherited pre- and
postconditions are still ignored after the front-end phase, this will be changed
next.

The base library had to be changed at many places to use `else` or `then` in
their contracts, and contracts that are redundant with the inherited contract
have been removed.

To avoid poor performance when inheriting index checks from `Sequence`, added a
new feature `Sequence.is_valid_index` that is redefined by 'array`.

Added `check` where a precondition would otherwise have to be strengthened:
shift operation for `uint`.

Added rules for `redef`, and contracts with `else`/`then` to reference manual.