ast: Do not create error for auto-generated fields in choice
This problem occurs in fuzion-lang.dev's design/examples/red_choice.fz example,
were subsequent errors are produced for the `result` field of a choice that
contains code.

This patch first changes the order in which errors are produces such that an
error for a choice that contains code is produced first. Then, any subsequent
errors for artificial fields in a choice are suppressed.