ast: Fix post condition for constructors
The result of a constructor is the new instance, so we need to pass `Current` as
the result argument.

For this to be possible, the type of the result in the post condition feature
must be `f.thisType()`, ie. exactly `f` and not `f.this`.

This fixes current failures in fuzion-lang.dev's tutorial examples like
`modify_array1.fz`.

This does not work for post conditions in constructors that are called in an
inherits clause, will create an issue for this.