util: Supress any further error output during shutdown due to errors, fix #3142
The problem was that `Errors.exit` throws a `FatalError`, which means that the
synchronized methods for error output are exited and any other thread may
produce error output during their own shutdown.

Now, a flag supresses further error output once we have started shutting down
due to errors.

To avoid running into this problem again this patch removes `Errors.exit` and
instead sets the `_shutting_down_ ` flag in `showAndExit`, which is synchronized
and combines the final error statistic reporting with shutting down.

This has the side-effect that the error statistics are also shown in case of a
fatal error, so some expected error output has to be updated.