lib: Add fallible effect
fallible is an template for effects that could abort an operation with a
failure.  Fallible has a type parameter for the error that is passed to `cause`,
which is the feature resulting in an abort.

This permits code to handle, e.g., a postcondition failure, via

  res := fuzion.runtime.post_fault
           .try
             do something that may to `fuzion.runtime.post_fault.env.cause`
           .catch s->
             handle error `s`

Note that this permits a hierachy of `fallible` effects, e.g., `post_fault.try`
will catch only post-condition failures, while `fault.try` will catch post
condition failures and other failures such as `check` or, eventually,
pre-conditions, etc.