parser: use keyword (for/do/while/until) as reference for indentation
this also _fixes_ the surprise loop of:
```
for i := 0, i+1
until i > 3 do
  say "hey!"
```

Here the `do` does not belong to the first loop. Instead it starts a new - infinite - loop.
fz then complains that `say "hey"!` is indented which it should not. Previously this say was considered to be part of the `do`-loop.

fixes #3119