ast: Workaround for #3160 for nested and inherited type features
The problem is that inheriting nested type features may result in outer type
features not getting their type parameters set correctly to the actual types.
This is ugly but most likely harmless since the outer types are unit types
and their features will not be called via the broken outer type.

This will still need a proper fix, but I would like to get this out of the way
for now such that #3157 and the work on effects for pre-/post-conditions is not
blocked.