dfa/tests: Reduce DFA accuracy to avoid state explosion for `String.upper_case`
The massive number of codepoint values created by
`character_encodings.unicode.data` cause the DFA to run forever. This patch
joins instances created in the same clazz to avoid this.

This might result in inaccuracy, e.g., the code

  a(p bool) is
  i1 := a false
  i2 := a true
  v := i1.p
  if v
    panic "unreachable"

will have the panic effect, while the code

  a(p bool) is
  i1 => a false
  i2 := a true
  v := i1.p
  if v
    panic "unreachable"

will not since the two instance of `a` are created in different clazzes `i1` and
`universe`, while the previous example creates both within `universe`.

This change permits to re-enable tests/unicode.