lib: Add default implementation of numeric operator preconditions
Since preconditions are inherited now implementations of these are needed by all
children of numeric. These implementations are missing, e.g., in fraction.

To avoid requiring adding all of these, this patch provides default
implementations that are just `true`, so children are supposed to implement
operations that are complete and redefine these for partial operations.

This fixes the abstract feature error produces by examples using fractions.