lib: make numeric operations preconditions depend on `debug`, not `safety`
A numeric overflow does not by itself result in unsafe operation (see Java where
ignoring overflows is the default behaviour), so they do not need to be checked
if `debug`ging is disabled.

This fixes fuzion-lang.dev's `tutorial/examles/fiball.fz` that stopped working
now this precondition inheritance is supported.