ast: fix outer for actual arg type inference used with partial application
A formal argument type that is inferred from the actual argument requires the
resolution of the type of the actual argument, which requires the outer feature
at the call site.  This outer feature can change as a result of partial
application, lazy evaluation or lambdas, so we must use the actual outer type.

This fixes problems due to the original outer and the final outer not being the
same by no longer recording the outer with the initial call but by using that
last outer type the call was resolved against.