ast/tests: Add type resolution for lambda to fix #3315 fix #3308
The type resolution was missing for arguments in a partial call.

The type feature is required as well and is created explicitly, marked this with
NYI since it is unclear why this does not happen automatically.

The issue from #3315 is no longer commented out in tests/reg_issue3308 and an
even simpler example was added as well.