lib: Make `panic` and `try` inherit from `eff.fallible`
`panic` is the standard `fallible` for quick-and-dirty handling of a severe
error (like a programming bug), while `try` is the standard fallible for an
expected error (like an I/O error).

`option.this` can apparently be used as type parameter in inheritance call of
`option`, so remove the NYI-comments and commented code.

Also add precondition to `abort` to ensure that effect is intalled.