ast: When checking actual type par against constraints, do not walk through inh chain of choice Fix #3362
Since choice types cannot be called, the result of a call to `choice X Y Z` is
`void`, which lead to choice types that inherit from choice to be accepted to
match any constraint. This patch fixes this by not walking through the parents
if the actual type parameter is a choice.

This also adds a regression test with the original code from #3362 and a minimalistic example.