air/dfa: Move reporting missing implementations of abstracts to DFA, fix #3359
The air phase is not accurate enough, which is why these errors were disabled
for type features.  Reporting them all during DFA avoids this and we can handle
missing abstract type features the same way as norml abstract features.

The code for handling this is still a bit spread out ofer packages air, fuir and
dfa, a NYI comment has been added to fix this when air and dfa phases are
joined.

Also adds a regression test.