dfa: do not record intrinsics as escaping
fixes #3429

Problem was that loop in `array.type.new` was not detected to be eligible for TCO, so we got a stack overflow:

```
    internal := fuzion.sys.internal_array_init T length
    for x in indices do
      internal[x] := init x
```

Reason was that call to `internal_array.setel` (and all other intrinsics) -  and thus also `internal[x] :=` -  were marked as escaping.

I believe it should currently not be possible for an instance to escape via intrinsic, so I just excluded calls to intrinsics from being marked as escaping.