lib: Remove `Types`, move `Types.get` to `type_as_value`, fix #3431
This removes the old-style type-feature `Types`.  The only inner feature. `get`,
unfortunately cannot be moved to `Type.type.get` or any other type feature, see
#3431 for details. Instead, `get` gets moved to `universe.type_as_value`.

Also added `Any.type_value` as an alternative. This would be a much better place
for `type_as_value`, but it cannot be within a type feature.

In any case, the following three expressions are equivalent

   abc.type
   type_as_value abc
   abc.type_value

the preferred version is the first.

Note that this patch also contains a fix in
`AbstractFeature.existingOrNewTypeFeature` that was triggered by adding a type
feature to `Any.fz`.