ast/tests: join precondition constraint handling to Context
So the handling of `pre T : x` and `if T : x` is at the same place.  This should
eventually make it possible to reduce the pair `outer`/`context` to a single
parameter `context`.

This fixes handling of complex `if T : x` constraint propagation into nested
lambda expressions, added test cases for this.