lib: support `ref` effect types, make all effects simple, remove `effect_mode`
Add a new term for adding an effect instance to the current context and runing
code: to `instate` an effect.

Effects are identified by their type. To support dynamic binding for operations
on an effect, we must distinguish the type used to identy the effect, and the
type of the actual effect value.  A new feature `effect.type.instate` permits
this as follows: Say we have a `ref` effedt `color` with two instances types
`green` and `purple`:

    color ref : effect is
      get String => abstract

    green  : color is
      redef get => "green"
    purple : color is
      redef get => "purple"

To `instate` one of them, we need to call `instate` on the effect type `color`,
providing an argument of the value type `green` or `purple`:

    color.instate green  ()->{say "color is {color.env.get}"}
    color.instate purple ()->{say "color is {color.env.get}"}

This avoids the need of using effect handlers, e.g. `io.out` currently uses an
argument `Print_Handler`, while a `ref` effect `io.Out` could have children like
`stdout` that directly redefines `io.Out`'s operations.

This patch does a major cleanup in `effect.fz`, in particular

- effect intrinsics are now all type features of `effect, they are `module`
  visible and have names ending `0`.

- two features `effect.type.instate` permit instating an effect with or without
  support for a default value in case of an abort.

- `effect.instate_self` is a convenience function similar to `effect.run`, which
  should be removed. `instate_self` does not work for inherited effect
  values instated for `ref` effect types.

- new featue `effect.type.default` to instate a default effect value.

- new featues `effect.type.replace` and `effect.replace` replace an instated
  effect by a new value. They replace `effect_mode.replace`

- new feature `effect.type.abort` to replace and abort an instated effect.

- `simple_effect` is no longer needed and marked for removal using `NYI:
  CLEANUP`

- Marked `use`, `go`. `run` with `NYI: CLEANUP` since these should be removed
  and replaced by calls to `instate` or `instate_self`.

- tests were changed to no longer use `simple_effect`, `use`, `go`, or `run`.