ast: In call, relax target type lookup to allow for partial application, fix #3658
The target of an `infix` call may not be found in case partial application will
later turn this into a lambda.  For this to work, the lookup of the type of the
target must be lazy using `Expr.typeForInferencing()` instead of `Expr.type()`.

This patch uses `Expr.typeForInferencing()` in `Call.targetTypeOrConstraint()`
and returns `Types.t_UNDEFINED` in case there is some problem obtaining this.
This is then further propagated as a `null` result for `Call.targetFeature()`.

Furthermore, calls to `reportPendingError` were removed, they are not performed
later during `checkTypes`.