ast: Fix resolution of actuals after partial application, fix #3674
Partial applicaton may change the list of arguments to a call, so when resolving
these arguments, we must use actuals list that was possibly changed.

For some reason that apparently has been fixed by now, we had to create the
iterator earlier.  This is no longer needed and even harmful for partial
application.

This patch also explicitly resets `_calledFeature` to null in case of partial
application to stop any pending resolution of actual arguments.

Also added a regression test.