c: support for `effect.finally` to be called if surrounding effect aborts
For this, there is no longer a thread local jump buffer for each effect clazz,
but only one jump buffer for any effect abort.  The value passed to longjump and
returned by setjmp is the effect id.

On an instate, the id returned from setjump is then used to decide if this is a
local abort, such that instate woudl return the result of call_def, or we have
an abort of a surrounding effect, i.e., we have to longjump further to the next
instate.