ast: improve logic for order of duplicate feature declaration errors
Some Jenkins tests show messages reporting the earlier feature as the culprit,
which is strange. The new logic blames the feature that is declared in currently
compiled source code, and if both are, the one that is declared later (or, in
the source file with a name that is alphabetically behind the other).