c: remove uses of `FUIR.isIntrinsicUsed`
This can be done fully be the backend, no need to have this in FUIR.  And the
FUIR implementation always returned `true`, so it did not work well anyway.

Note that this patch does not remove `FUIR.isIntrinsicUsed`. I want to do this
as part of PR #3820 and avoid merge conflicts that would result from changing
FUIR now.