fuir: Fix precondition failure for `clazzResultClazz` on choice clazzasref
This precondition was just added and it fails in the C test run of
tests/base32. The problem is that `clazzResultClazz` was used to obtain the
result clazz or an argument, but that argument was obtained by `clazzArgClazz`,
which is already the result clazz.  In this test, the result clazz happens to be
a choice such that this bug was found.